requirements.txt
asgiref==3.4.1
azure-ai-textanalytics==5.2.0b1
azure-common==1.1.27
azure-core==1.19.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7